chore(core): move `isAuthStore` type guard to new `asserters` module
