+++ packages/sanity/src/core/store/_legacy/authStore/utils/asserters.ts
@@ -1,4 +1,23 @@
+import type {AuthStore} from '../types'
 import type {LoginMethod, CookielessCompatibleLoginMethod} from '../../../../config'
+import {isRecord} from '../../../../util'
+
+/**
+ * Duck-type check for whether or not this looks like an auth store
+ *
+ * @param maybeStore - Item to check if matches the AuthStore interface
+ * @returns True if auth store, false otherwise
+ * @internal
+ */
+export function isAuthStore(maybeStore: unknown): maybeStore is AuthStore {
+  return (
+    isRecord(maybeStore) &&
+    'state' in maybeStore &&
+    isRecord(maybeStore.state) &&
+    'subscribe' in maybeStore.state &&
+    typeof maybeStore.state.subscribe === 'function'
+  )
+}
 
 /**
  * Check whether the provided login method is compatible with cookieless auth, e.g. whether any
